Organize the groups flow files.
