+++ src/main/Start.java
@@ -10,19 +10,20 @@ import common.Comm;
 import common.Configurations;
 import common.Logger_;
 import common.structures.ConfigStructure;
+import common.structures.GroupStructure;
 import common.structures.ToPost;
 import impl.CheckNotifications;
 import impl.DoLogin;
 import impl.DoLogout;
 import impl.FileXML;
-import impl.GoToGroups;
 import impl.GoToManagePages;
-import impl.ManageGroupsListFiles;
 import impl.ManagePagesFiles;
 import impl.PagePostsFiles;
 import impl.Posts;
 import impl.group.ManageFiles;
 import impl.group.Publish;
+import impl.groups.GoToGroups;
+import impl.groups.ManageGroupsListFiles;
 
 public class Start {
